requirements.txt
dhanhq==1.3.6
python-telegram-bot==20.7
requests==2.31.0
asyncio==3.4.3
